pyproject.toml
[project]
name = "cua-operator"
description = "Computer-Using Agent (CUA) Operator"
version = "0.0.0"
requires-python = ">=3.8"
readme = "README.md"
license-files = ["LICENSE"]
dependencies = [
  "pillow"
]

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio", "openai"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["cua_operator"]
